move channel serialization to dedicated package
