+++ modules/core/src/commonMain/kotlin/fr/acinq/lightning/serialization/channel/Serialization.kt
@@ -1,4 +1,4 @@
-package fr.acinq.lightning.serialization
+package fr.acinq.lightning.serialization.channel
 
 import fr.acinq.bitcoin.crypto.Pack
 import fr.acinq.lightning.channel.states.PersistedChannelState
@@ -6,13 +6,13 @@ import fr.acinq.lightning.channel.states.PersistedChannelState
 object Serialization {
 
     fun serialize(state: PersistedChannelState): ByteArray {
-        return fr.acinq.lightning.serialization.v4.Serialization.serialize(state)
+        return fr.acinq.lightning.serialization.channel.v4.Serialization.serialize(state)
     }
 
     fun deserialize(bin: ByteArray): DeserializationResult {
         return when {
             // v4 uses a 1-byte version discriminator
-            bin[0].toInt() == 4 -> DeserializationResult.Success(fr.acinq.lightning.serialization.v4.Deserialization.deserialize(bin))
+            bin[0].toInt() == 4 -> DeserializationResult.Success(fr.acinq.lightning.serialization.channel.v4.Deserialization.deserialize(bin))
             // v2/v3 use a 4-bytes version discriminator
             Pack.int32BE(bin) == 3 -> DeserializationResult.Success(fr.acinq.lightning.serialization.v3.Serialization.deserialize(bin))
             Pack.int32BE(bin) == 2 -> DeserializationResult.Success(fr.acinq.lightning.serialization.v2.Serialization.deserialize(bin))
