+++ modules/core/src/commonMain/kotlin/fr/acinq/lightning/serialization/channel/v2/ChannelState.kt
