+++ modules/core/src/commonMain/kotlin/fr/acinq/lightning/serialization/channel/v3/Serialization.kt
