+++ modules/core/src/commonTest/kotlin/fr/acinq/lightning/utils/Extensions.kt
@@ -1,7 +1,7 @@
 package fr.acinq.lightning.utils
 
 import fr.acinq.lightning.channel.states.PersistedChannelState
-import fr.acinq.lightning.serialization.Serialization
+import fr.acinq.lightning.serialization.channel.Serialization
 
 val Serialization.DeserializationResult.value: PersistedChannelState
     get() = (this as Serialization.DeserializationResult.Success).state
